+++ src/Configuration/Loader/ArrayConfigLoaderFactory.php
@@ -10,14 +10,15 @@ use ArtARTs36\MergeRequestLinter\Condition\Operator\OperatorFactory;
 use ArtARTs36\MergeRequestLinter\Condition\Operator\OperatorResolver;
 use ArtARTs36\MergeRequestLinter\Configuration\ConfigFormat;
 use ArtARTs36\MergeRequestLinter\Configuration\Loader\Loaders\ArrayLoader;
-use ArtARTs36\MergeRequestLinter\Configuration\Loader\Mapper\CredentialMapper;
 use ArtARTs36\MergeRequestLinter\Configuration\Loader\Mapper\ArrayConfigHydrator;
+use ArtARTs36\MergeRequestLinter\Configuration\Loader\Mapper\CredentialMapper;
 use ArtARTs36\MergeRequestLinter\Configuration\Loader\Mapper\RulesMapper;
 use ArtARTs36\MergeRequestLinter\Configuration\Value\EnvTransformer;
 use ArtARTs36\MergeRequestLinter\Configuration\Value\FileTransformer;
 use ArtARTs36\MergeRequestLinter\Contracts\Config\ConfigLoader;
 use ArtARTs36\MergeRequestLinter\Contracts\Environment\Environment;
 use ArtARTs36\MergeRequestLinter\Contracts\Report\MetricManager;
+use ArtARTs36\MergeRequestLinter\Infrastructure\Container\MapContainer;
 use ArtARTs36\MergeRequestLinter\Rule\CustomRule\OperatorRulesExecutor;
 use ArtARTs36\MergeRequestLinter\Rule\CustomRule\RulesExecutor;
 use ArtARTs36\MergeRequestLinter\Rule\DefaultRules;
@@ -27,7 +28,6 @@ use ArtARTs36\MergeRequestLinter\Rule\Factory\ConditionRuleFactory;
 use ArtARTs36\MergeRequestLinter\Rule\Factory\Constructor\ConstructorFinder;
 use ArtARTs36\MergeRequestLinter\Rule\Factory\Resolver;
 use ArtARTs36\MergeRequestLinter\Rule\Factory\RuleFactory;
-use ArtARTs36\MergeRequestLinter\Support\MapContainer;
 use ArtARTs36\MergeRequestLinter\Support\Reflector\CallbackPropertyExtractor;
 use ArtARTs36\MergeRequestLinter\Support\Text\DecoderFactory;
