+++ src/Console/Application/ApplicationFactory.php
@@ -8,8 +8,8 @@ use ArtARTs36\MergeRequestLinter\Configuration\ConfigFormat;
 use ArtARTs36\MergeRequestLinter\Configuration\Copier;
 use ArtARTs36\MergeRequestLinter\Configuration\Loader\ArrayConfigLoaderFactory;
 use ArtARTs36\MergeRequestLinter\Configuration\Loader\Loaders\CompositeLoader;
-use ArtARTs36\MergeRequestLinter\Configuration\Loader\Loaders\Proxy;
 use ArtARTs36\MergeRequestLinter\Configuration\Loader\Loaders\PhpLoader;
+use ArtARTs36\MergeRequestLinter\Configuration\Loader\Loaders\Proxy;
 use ArtARTs36\MergeRequestLinter\Configuration\Resolver\ConfigResolver;
 use ArtARTs36\MergeRequestLinter\Configuration\Resolver\MetricableConfigResolver;
 use ArtARTs36\MergeRequestLinter\Configuration\Resolver\PathResolver;
@@ -18,13 +18,13 @@ use ArtARTs36\MergeRequestLinter\Console\Command\InfoCommand;
 use ArtARTs36\MergeRequestLinter\Console\Command\InstallCommand;
 use ArtARTs36\MergeRequestLinter\Console\Command\LintCommand;
 use ArtARTs36\MergeRequestLinter\Environment\LocalEnvironment;
+use ArtARTs36\MergeRequestLinter\Infrastructure\Container\MapContainer;
 use ArtARTs36\MergeRequestLinter\IO\Console\ConsoleLoggerFactory;
 use ArtARTs36\MergeRequestLinter\Linter\Runner\RunnerFactory as LinterRunnerFactory;
 use ArtARTs36\MergeRequestLinter\Report\Metrics\Manager\MemoryMetricManager;
 use ArtARTs36\MergeRequestLinter\Rule\Dumper\RuleDumper;
 use ArtARTs36\MergeRequestLinter\Rule\Factory\Argument\ArgumentResolverFactory;
 use ArtARTs36\MergeRequestLinter\Support\File\Directory;
-use ArtARTs36\MergeRequestLinter\Support\MapContainer;
 use ArtARTs36\MergeRequestLinter\Support\ToolInfo\ToolInfoFactory;
 use Symfony\Component\Console\Output\OutputInterface;
