+++ src/Infrastructure/Container/MapContainer.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace ArtARTs36\MergeRequestLinter\Support;
+namespace ArtARTs36\MergeRequestLinter\Infrastructure\Container;
 
 use ArtARTs36\MergeRequestLinter\Exception\MergeRequestLinterException;
 use Psr\Container\ContainerInterface;
