+++ src/Infrastructure/Http/MetricableClient.php
@@ -1,10 +1,11 @@
 <?php
 
-namespace ArtARTs36\MergeRequestLinter\Support\Http;
+namespace ArtARTs36\MergeRequestLinter\Infrastructure\Http;
 
 use ArtARTs36\MergeRequestLinter\Contracts\HTTP\Client;
 use ArtARTs36\MergeRequestLinter\Contracts\Report\MetricManager;
 use ArtARTs36\MergeRequestLinter\Report\Metrics\MetricSubject;
+use ArtARTs36\MergeRequestLinter\Support\DataStructure\Arrayee;
 use ArtARTs36\MergeRequestLinter\Support\Time\Timer;
 use Psr\Http\Message\RequestInterface;
 use Psr\Http\Message\ResponseInterface;
@@ -38,7 +39,7 @@ class MetricableClient implements Client
 
         $responses = $this->client->sendAsyncRequests($requests);
 
-        $hosts = implode(', ', $this->getHosts($requests));
+        $hosts = $this->getHosts($requests)->implode(', ');
 
         $this->metrics->add(new MetricSubject(
             'http_send_request',
@@ -50,9 +51,9 @@ class MetricableClient implements Client
 
     /**
      * @param array<RequestInterface> $requests
-     * @return array<string>
+     * @return Arrayee<int, string>
      */
-    private function getHosts(array $requests): array
+    private function getHosts(array $requests): Arrayee
     {
         $hosts = [];
 
@@ -60,6 +61,6 @@ class MetricableClient implements Client
             $hosts[$request->getUri()->getHost()] = $request->getUri()->getHost();
         }
 
-        return $hosts;
+        return new Arrayee(array_values($hosts));
     }
 }
