+++ src/Linter/Runner/RunnerFactory.php
@@ -5,13 +5,13 @@ namespace ArtARTs36\MergeRequestLinter\Linter\Runner;
 use ArtARTs36\MergeRequestLinter\CI\System\SystemFactory;
 use ArtARTs36\MergeRequestLinter\Configuration\Config;
 use ArtARTs36\MergeRequestLinter\Contracts\CI\CiSystem;
+use ArtARTs36\MergeRequestLinter\Contracts\DataStructure\Map;
 use ArtARTs36\MergeRequestLinter\Contracts\Environment\Environment;
 use ArtARTs36\MergeRequestLinter\Contracts\Linter\LinterRunner;
 use ArtARTs36\MergeRequestLinter\Contracts\Linter\LinterRunnerFactory;
 use ArtARTs36\MergeRequestLinter\Contracts\Report\MetricManager;
+use ArtARTs36\MergeRequestLinter\Infrastructure\Http\ClientFactory;
 use ArtARTs36\MergeRequestLinter\Request\Fetcher\CiRequestFetcher;
-use ArtARTs36\MergeRequestLinter\Contracts\DataStructure\Map;
-use ArtARTs36\MergeRequestLinter\Support\Http\ClientFactory;
 use Psr\Log\LoggerInterface;
 
 class RunnerFactory implements LinterRunnerFactory
