+++ src/Support/ToolInfo/ToolInfoFactory.php
@@ -5,8 +5,8 @@ namespace ArtARTs36\MergeRequestLinter\Support\ToolInfo;
 use ArtARTs36\MergeRequestLinter\CI\Credentials\Token;
 use ArtARTs36\MergeRequestLinter\CI\System\Github\GraphQL\Client;
 use ArtARTs36\MergeRequestLinter\CI\System\Github\GraphQL\PullRequest\PullRequestSchema;
+use ArtARTs36\MergeRequestLinter\Infrastructure\Http\ClientGuzzleWrapper;
 use ArtARTs36\MergeRequestLinter\Request\Data\Diff\DiffMapper;
-use ArtARTs36\MergeRequestLinter\Support\Http\ClientGuzzleWrapper;
 use Psr\Log\NullLogger;
 
 class ToolInfoFactory
